fix: fix bug in mockData, rename file
