+++ src/components/visualization/BarChart.tsx
